Added support for Hibernate 5 and 5.1.
For simplicity in the code, this is compiling against Hibernate 5.0.7
instead of 5.1.0. The tests have been run against both 5.0.7 and 5.1.0
and pass, but it's simpler to compile against 5.0.7 because CacheEntry
still has the areLazyPropertiesUnfetched() method. In order to compile
against 5.1.0, the Hibernate5CacheEntrySerializer would need to be
reimplemented serialize entirely via Reflection in order to capture
the areLazyPropertiesUnfetched() value.

- Essentially duplicates hazelcast-hibernate4, with API/SPI updates
  for Hibernate 5
- Unlike Hibernate 4, on Hibernate 5 there are different cache key
  types for collection/entity and natural ID, so two serializers
  are required
  - The CacheKey class has been renamed to OldCacheKeyImplementation,
    and is now package-private, so serialization must be done entirely
    via Reflection
  - The StreamSerializer implementations for both of the key types
    are StreamSerializer<Object> because the types they're bound to
    are package-private